feat: implement default admin user
