+++ src/scripts/data/default_admin.json
@@ -1,0 +1,6 @@
+{
+	"name": { "first": "DJI", "last": "Admin" },
+	"email": "admin@dji.com",
+	"phone": "+974 1234 5678",
+	"password": "9cda85dc79aa1ce43a12a239575cec488e87b4d877cfaa31079c0e8b4b9c1569"
+}
