+++ src/scripts/index.ts
@@ -1,5 +1,4 @@
 //	Package Imports
-import Cookies from 'js-cookie';
 import './icons';
 
 //	Component Imports
@@ -15,13 +14,12 @@ import '@/components/TextInput/logic';
 //	Repository Imports
 import { UserRepository } from '@/scripts/db/UserRepository';
 
+//	Set the number of users to the number of users in the local storage whenever the local storage is updated
+UserRepository.updateUsersList();
+window.addEventListener('storage', UserRepository.updateUsersList);
+
+//	If the default admin is not in the local storage, add it
+if (!UserRepository.getUser(0)) UserRepository.addDefaultAdmin();
+
 //	Get the current user from the cookies and store it in the window object
-const userId = Cookies.get('user');
-if (userId) {
-	try {
-		const user = UserRepository.loginUserById(Number(userId));
-		window.currentUser = user;
-	} catch (error) {
-		Cookies.remove('user');
-	}
-}
+UserRepository.loginFromCookies();
